Display crate specific links
We always a link to docs.rs instead of the specified key because why not?
The `documentation = "URL"` item was introduced before docs.rs was born.
Nowadays, docs.rs is all you need! Maybe in the future we could build a
"local docs.rs service" that creates documentation for crates on-demand.