Optimize html rendering for formatted code
Applies 3 different optimisations to html formatted
output, reducing the total amount of spent chars
for the formatted html:

1. It merges consecutive spans with the same "style".
2. It merges consecutive spans if the second span only
   consists of whitespace characters.
3. It does all these three things across lines

For this, we have copied a bit of the syntect source
code into three new files.

The resulting character counts for an example snippet
from the piston readme (crate version 0.35.0) are:

6196 : State before this commit, without optimisations
3388 : Only optimisation 1 applied
3136 : Optimisations 1 and 2 applied
2704 : Optimisations 1, 2 and 3 applied

So we've achieved savings of 56%!

Results for the html pages we serve for real world
crates that have some snippets (only the html pages):

Before the patch:

 crate  # vers   # compressed # uncompressed
 ############################################
 rand   # 0.35.0 # 2.69 KB    # 13.42 KB
 piston # 0.3.18 # 3.31 KB    # 13.97 KB

After the patch:

 crate  # vers   # compressed # uncompressed
 ############################################
 rand   # 0.35.0 # 2.60 KB    # 8.43 KB
 piston # 0.3.18 # 3.23 KB    # 10.56 KB

Improvement:

 crate  # vers   # compressed # uncompressed
 ############################################
 rand   # 0.35.0 # 3%         # 37%
 piston # 0.3.18 # 2%         # 24%

So there is an improvement even after compression,
although it is only very little.
Before compression, the improvement is huge
however! And don't forget, this is what the browser
has to parse, keep in memory, and run its
algorithms over.

Optimisations 1 and 2 got upstreamed to syntect:
https://github.com/trishume/syntect/pull/123

There are possible further optimisations:

* Counting the number of the most used styling setting
  and setting it as the default, so that stuff in that
  style doesn't have to be put into <span> tags.
* A refined version of above trick: Compute the most
  often used style of a stack of sliding windows of
  different lengths and if there is a different most
  used styling setting at some level than at the level
  above, set it to that. This way we could build a
  "tree" of styling options.
  However, this will most likely only help with a
  subset of workloads and is a bit involved.
* Using <span class="..."> instead of inline CSS. The
  basic idea is to derive the set of possible resulting
  Style values for a given theme and giving each one an
  id. Then we could serve the CSS separately and adopt
  a stricter CSP (which is always desireable).