Add a "most versions" ranking to the index page
This metric is utter shit for basically anything,
but at least we can calculate it easily,
and displaying it is better than a void.

Yes, this ranking might influence some people to
game it, so that they release more often. This
of course runs contrary to any mirroring efforts
as many versions of the same crate bloat mirrors.