Make MutCrateSource the default
Non mutable crate sources are nice but most people will want to be
generic over mutable and non mutable crate sources anyway.
Having a non mutable CrateSource around as a separate trait might be
better for having minimal requirements but generally isn't really
useful. If it is worth the extra overhead, it can be readded.